+++ driver/nativeapi/mysql_native_connection_wrapper.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2009, 2019, Oracle and/or its affiliates. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License, version 2.0, as
@@ -383,7 +383,7 @@ MySQL_NativeConnectionWrapper::get_option(::sql::mysql::MySQL_Connection_Options
 int
 MySQL_NativeConnectionWrapper::query(const SQLString & stmt_str)
 {
-  return api->real_query(mysql, stmt_str.c_str(), stmt_str.length());
+  return api->real_query(mysql, stmt_str.c_str(), static_cast<unsigned long>(stmt_str.length()));
 }
 /* }}} */
